♻️ getDiscount 분리
