+++ src/basic/shared/lib/date/config.js
@@ -1,0 +1,3 @@
+export const DAYS = Object.freeze({
+  TUESDAY: 2,
+});
