+++ src/basic/shared/lib/date/index.js
@@ -1,0 +1,3 @@
+import { DAYS } from './config.js';
+
+export const isTuesday = () => new Date().getDay() === DAYS.TUESDAY;
